requirements.txt
pandas>=2.0
numpy>=1.23
scikit-learn>=1.2
pyyaml>=6.0
matplotlib>=3.7
